// src/components/pdi/pdi-types.ts
export interface DamageMarker {
    id: string
    x: number  // X coordinate (percentage 0-100)
    y: number  // Y coordinate (percentage 0-100)
    type: DamageType
    code: DamageCode  // Short code for the damage type
    severity: DamageSeverity
    description?: string
    view: DamageView  // Which vehicle view this damage is on
}

// Exterior damage codes: D (Dent), S (Scratch), CH (Chip), CR (Crack)
// Interior damage codes: TR (Tear), ST (Stain), SC (Scratch), BR (Broken), NW (Not Working)
export type DamageCode =
    | 'D'    // Dent
    | 'S'    // Scratch
    | 'CH'   // Chip
    | 'CR'   // Crack
    | 'TR'   // Tear (Interior)
    | 'ST'   // Stain (Interior)
    | 'SC'   // Scratch (Interior)
    | 'BR'   // Broken (Interior)
    | 'NW'   // Not Working (Interior)
    | 'RS'   // Rust
    | 'MS'   // Missing
    | 'OT'   // Other

export type DamageType =
    | 'scratch'
    | 'dent'
    | 'crack'
    | 'chip'
    | 'rust'
    | 'paint-damage'
    | 'broken'
    | 'missing'
    | 'tear'
    | 'stain'
    | 'not-working'
    | 'other'

export type DamageSeverity = 'minor' | 'moderate' | 'major'

export type DamageView = 'top' | 'side' | 'interior' | 'boot'


export interface VehicleDamageData {
    markers: DamageMarker[]
    notes?: string
}